feat: refactor utility imports to appUtils and remove old utils.js
