+++ src/js/calc/userDefined.js
@@ -2,7 +2,7 @@ import { dom } from '../dom'
 import { refreshEditor, udfInput, uduInput } from '../editor/editor'
 import { calculate, math } from '../eval'
 import { modal, showError } from '../ui/modal'
-import { app, store } from '../utils'
+import { app, store } from '../appUtils'
 import { DateTime as luxon } from 'luxon'
 import * as formulajs from '@formulajs/formulajs'
 import nerdamer from 'nerdamer-prime/all.js'
