+++ src/js/ui/pages.js
@@ -2,7 +2,7 @@ import { dom } from '../dom'
 import { cm } from '../editor/editor'
 import { calculate, renderAnswersToHTML, syncOutputHeights } from '../eval'
 import { confirm, modal, notify } from './modal'
-import { app, isElectron, store } from '../utils'
+import { app, isElectron, store } from '../appUtils'
 import { escapeHTML } from '../core/coreUtils.js'
 import { syncPageSave, syncPageRename, syncPageDelete } from '../sync'
